Add files via upload
A temperature converter is a simple web application that allows users to convert temperatures between different units, such as Celsius, Fahrenheit, and Kelvin. It can be built using HTML, CSS, and JavaScript. A temperature converter can be useful for learning about the formulas and calculations involved in temperature conversion, as well as for practical purposes such as checking the weather or cooking.


Temperature Converter
This is a temperature converter web application that I built using HTML, CSS, and JavaScript. It allows users to convert temperatures between Celsius, Fahrenheit, and Kelvin units. It also displays the corresponding color of the temperature on the background.

Technologies and Tools
HTML: A markup language for creating web pages
CSS: A style sheet language for describing the presentation of web pages
JavaScript: A scripting language for adding interactivity and functionality to web pages
VS Code: A code editor for developing and debugging web applications
Git: A version control system for tracking changes in source code
GitHub: A platform for hosting and collaborating on code projects
Features and Functionality
Input: Users can enter a temperature value in any unit (Celsius, Fahrenheit, or Kelvin) in the input field. The input field will validate the input and only accept numerical values.
Conversion: Users can click on the convert button to convert the input temperature to the other two units. The converted temperatures will be displayed in the output fields below the input field.
Color: Users can see the color of the input temperature on the background of the web page. The color will change according to the input value and unit. For example, if the input is 0°C, the background will be blue; if the input is 100°F, the background will be red; if the input is 273.15 K, the background will be white.
Reset: Users can click on the reset button to clear the input and output fields and reset the background color to black.
Installation and Usage
To run this project locally, you need to have a code editor such as VS Code installed on your machine.

Clone this repository or download the ZIP file.
Open the project folder in your code editor.
Open the index.html file in your browser or use a live server extension to launch the web application.
Enter a temperature value in any unit in the input field and click on the convert button to see the results.